rename trip to ride

+++ backend/database/migrations/2024_09_08_204060_create_driver_locations_table.php
@@ -8,20 +8,27 @@ return new class extends Migration
 {
     public function up(): void
     {
-        Schema::create('driver_locations', function (Blueprint $table) {
-            $table->id();
-            $table->foreignId('driver_id')->constrained()->onDelete('cascade');
-            $table->point('location')->nullable();
-            $table->timestamp('updated_at');
-            $table->foreignId('current_ride_id')->nullable()->constrained('rides')->nullOnDelete();
-            
-            // Add spatial index
-            $table->spatialIndex('location');
-        });
+        if (!Schema::hasTable('driver_locations')) {
+            Schema::create('driver_locations', function (Blueprint $table) {
+                $table->id();
+                $table->foreignId('driver_id')->constrained('drivers')->onDelete('cascade');
+                $table->point('location')->nullable();
+                $table->timestamp('updated_at');
+                $table->foreignId('current_ride_id')->nullable()->constrained('rides')->nullOnDelete();
+                $table->spatialIndex('location');
+            });
+        }
     }
 
     public function down(): void
     {
-        Schema::dropIfExists('driver_locations');
+        if (Schema::hasTable('driver_locations')) {
+            Schema::table('driver_locations', function (Blueprint $table) {
+                $table->dropForeign(['driver_id']);
+                $table->dropForeign(['current_ride_id']);
+            });
+    
+            Schema::dropIfExists('driver_locations');
+        }
     }
 };
